[refactor] Reduce code duplication
I could notice no slowdown and simple similar examples are fully
constevaluated by the compiler on godbolt. Because of the forceinline,
it should be no problem to call the non-template version from the
templated version of the same function and still profit from the speedup
from knowing the argument at compile-time.