[performance] Throw exceptions for very rare cases to reduce branches
This is the missing link! With this, the BitReader benchmark finally
yields similar results to the simple bit reading loop!

Note that completely removing the refillBuffer yields even more speedup
of ~50% from 1.2 GB/s to 1.7 GB/s when reading 16-bit strings!
However, I'm out of ideas how to pack this observation into the same
BitReader interface. When trying to use that, it would be necessary to
refill the buffer in an outer loop, i.e., from the call site.