[performance] Add prefetching strategy that can detect multiple streams
It is not used by default yet because it does not seem to improve much
over the previous one. I need to performance analyze this more in depth
but it seems like it prefetches much more than necessary, leading to a
lot of useless work being done.

   Parallelization                   : 12
   Hits
       Cache                         : 307497
       Prefetch Queue                : 405
   Misses                            : 280
   Hit Rate                          : 0.999091
   Prefetched Blocks                 : 4136

There only are ~600 blocks! Prefetching 4k should not be necessary.