[fix] Avoid 0x7FFFF000 (2'147'479'552) B limit for write syscall on Linux
Note that this limit should normally not be reached. It is only a
symptom of an underlying (performance) problem. Currently, it can be
triggered easily with a 2 GiB+ random file but even when adding
uncompressed block detection to the block finder, a gzip while using
only Fixed Huffman deflate blocks or non-zero padding for uncompressed
deflate blocks might still trigger this problem.