[refactor] MAX_CL_LENGTH is 15 not 16
Note that this is neither a fix nor really a performance optimization
because it only results in a very slightly smaller bit length frequency
map. On the contrary, it might introduce new bugs if I am missing
something. It improves readability because wrong but working values
might confuse code readers but it is not a documentation-only change
because it changes functionality. Ergo, it is a "refactor".