[doc] Profile access to shared file reader
Currently, we are reading 1.6x times the underlying file!
This is a performance bug introduced in:

    1001e4b [performance] Also look for uncompressed deflate blocks

The problem is that no uncompressed offsets are found leading to the
whole offset range being searched leading to frequent BitReader refills
and then we have to refill again when seeking back a large amount to
rerun the Dynamic Huffman deflate block finder!

Before that faulty commit, the read overhead of the file was only ~1%
instead of ~60%. Note that increasing the decode block size from 1 MiB
to 8 MiB also reduces the overhead to 5% probably because it finds some
false positive uncompressed block candidate and because the offset range
is much larger, the overhead for the uncompressed block finder is
smaller.

m pragzip && time src/tools/pragzip -P 0 -d -c 4GiB-base64.gz | wc -c

    [SharedFileReader::~SharedFileReader]
        seeks back    : (  0 <= 8e+06 +- 5e+06 <= 2.5e+07  ) B ( 20110  calls)
        seeks forward : (  0 <= 8e+06 +- 5e+06 <= 2.5e+07  ) B ( 19993  calls)
        reads         : (  0 <= 131100 +- 1000 <= 131100  ) B ( 40133  calls)
        read in total  5.25998e+09  B out of  3263906195  B,  i.e., read the file  1.61156  times

    Decompressed in total 4294967296 B in 4.20907 s -> 1020.41 MB/s
    4294967296

    real 0m4.223s