[doc] Update blockfinder benchmarks: 28.6 -> 30.0 MB/s
I'm not sure since when it has become faster. There are ~35 commits to
the last benchmarked commit.

I also tried a two-staged precode histogram lookup approach. First, a
lookup in a smaller (hopefully fitting in L1 cache) LUT of 4 KiB and
then the larger one with 4 MiB with 5 counts instead of only 3. However,
this degraded performance down to ~25 MB/s:

[findDeflateBlocksPragzipLUT with 13 bits] ( 25.07 <= 25.46 +- 0.17 <= 25.64 ) MB/s
[findDeflateBlocksPragzipLUT with 14 bits] ( 24.85 <= 25.2 +- 0.16 <= 25.36 ) MB/s
[findDeflateBlocksPragzipLUT with 15 bits] ( 25.29 <= 25.61 +- 0.13 <= 25.76 ) MB/s
[findDeflateBlocksPragzipLUT with 16 bits] ( 25.09 <= 25.44 +- 0.25 <= 25.8 ) MB/s
[findDeflateBlocksPragzipLUT with 17 bits] ( 25.04 <= 25.57 +- 0.24 <= 25.79 ) MB/s
[findDeflateBlocksPragzipLUT with 18 bits] ( 23.9 <= 24.9 +- 0.5 <= 25.3 ) MB/s