[fix] Distance code counts must be in 1-32 (0-31 stored value)!
I don't have a reproducing file and I'm still wondering how I could get
it wrong for so long and so often, duplicating the wrong threshold in
the test comments and such.

Fortunately, it seems to be rather rare. It should affect files with
lots of back-references and all possible distances.