[performance] Stop looking for blocks after 512 KiB to avoid slower than single-core performance for BGZF-like files: 100 -> 230 MB/s
The underlying problem is that the block finder only looks for non-final
blocks. Failing to find this should only be slightly slower than
single-core performance but, before this commit, the main thread would
see that the required offset is already being decompressed and waits for
it to finish. And only after it finishes and has failed to find anything
will a new decompression task with the correct offset be started.

Note that this also improves performance for files encoded with
pigz --huffman because in this case the block finder also fails!