[performance] Blockfetcher: Avoid waiting for non-existing blocks
The reason for this fix were thousands of mutex lock calls per
millisecond because it tries to prefetch blocks that are after the end
of the file.

I think the other reason was that this caused the block fetcher to wait
even for unused wrong prefetches. And those took long because they did
not stop looking until after the whole reange do decompress has been
analyze with the slow block finder.