[feature] Add command line option for toggling CRC32 computation
CRC32 computation is only implemented for --analyze and
pragzip -P 1. The Python bindings always use ParallelGzipReader
even with -P 1 and therefore is not affected.

Because CRC32 computation showed a bandwidth of ~4.6 GB/s,
the estimated impact is expected to be ~5%, which should
be negligible for the added safety it provides.

This 5% slowdown is also exactly what can be observed empirically:

m pragzip && src/tools/pragzip -P 1 -o /dev/null 4GiB-base64.gz

Before: Decompressed in total 4294967296 B:

    17.0846 s -> 251.394 MB/s
    17.1163 s -> 250.928 MB/s
    17.4955 s -> 245.49 MB/s
    17.0494 s -> 251.913 MB/s

After: Decompressed in total 4294967296 B in

    18.3190 s -> 234.454 MB/s
    18.2328 s -> 235.563 MB/s
    18.1222 s -> 237.001 MB/s
    17.9165 s -> 239.721 MB/s