[feature] Reduce maximum memory usage by 50x
The idea is simply: Just stop decompressing before the until offset.
This uses the property of the architecture that chunk offsets get
on-demand fetched in the worst case instead of throwing an error
in unexpected situations. It can worsen performance though because
of the intricate interplay with the prefetcher and block map, ...

The maximum decompressed chunk size is 20x the compressed chunk size.
This limits the main memory usage of the cache-prefetch architecture to
(20 * 4 MiB) * prefetch count * cache size, where prefetch count = P,
and cache size = std::max( P, 16 ). I.e., for large P, the memory
usage is limited to: 240 MiB * P. E.g. for P = 24: 5.625 GiB total.

Note that the cache size should be limited to 1 during pure sequential
decompression as is the case when decompressing with the CLI.