[doc] Add alternative name "rapidgzip" to main Readme
The alternative name is related to publishing and double-blind review.

Also, it seemed to sound better and conveys the design goal of a fast gzip decompression tool better but it is hard to change all usages of "pragzip" in already published content, and "pragzip" is shorter to write and pronounce. One outsider was wondering whether pragzip has anything to do with "pragmatic", an association I didn't intend. It also has nothing to do with Prague.

The prefix contains the same acronym "pra"/"rap" for "parallel random access" as pragzip and the "id" part could is a backronym for "indexed decompression" referring to the >optional< index files for speeding up random access and decompression even more.

In conclusion, I will start by adding the alternative spelling at selected locations to make it findable via Google instead of changing everything at once.