[fix] An exception was thrown when using --import-index with BGZF files
This got introduced when adding block splitting because it made it
possible for index boundaries to happen in the middle of deflate
streams. It isn't efficient for that to happen and should be fixed in a
separate commit to reduce the windows that need to be stored.

The bug that is fixed in this commit could also be triggered when trying
to import an index created by some other tool.