[style] ZlibInflateWrapper: Detect when there is not enough data to read the footer
This was not a bug because inflateReset2 and inflateGetHeader will
simply return Z_OK when called with avail_in == 0! Of course
gzipHeader.done will not be 1.
However, this would be a bug when trying to use this code with ISA-l
because ISA-l's semantic interface is slightly different and will return
an error if there is not enough data to read the footer.