[doc] Update rapidgzip timings via Python in Readme
In the future, the readme should instead show the scaling benchmarks
from the paper but redone with the improvements since then.

The benchmark results for decompression Silesia without an existing
index got slightly worse since the introduction of this commit:

    [feature] Reduce maximum memory usage by 50x

This is not too a bad tradeoff for the reduce memory usage.
I'm not entirely sure why performance got worse because of this.
Maybe the split chunks result in chunk cache evictions evictions?