[feature] GzipAnalyzer: Print back-reference length histogram
This is a preliminary study to gauge how much memory usage and copies
could be saved by keeping long back-references until the markers have
been resolves.

This would reduce the number of markers to replace and the memory
because unresolved markers are 16-bit instead of 8-bit symbols.

The tests and plots also give a guidance for a cut-off for
back-references to keep. Those of length 16 and higher seem to be good
candidates.