[performance] Avoid window recompressions for gztool indexes
src/tools/rapidgzip -v -P 1 --index-format gztool
    --export-index base64-512MiB.gz{.index,} 2>&1 | grep Index

    After  : 0.094111 s  0.09273 s  0.093241 s
    Before : 0.110074 s  0.10917 s  0.107952 s

The effect is not more pronounced because blocks on boundaries are not
compressed at all because they are on the critical path. In this case,
this leaves roughly half the windows that need to be compressed in the
first place because chunk splitting will split almost every chunk into
two.