[performance] Use sparsity information of window to increase compression
m rapidgzip && src/tools/rapidgzip -P 1 --index-format gztool \
    --export-index base64-512MiB.gz{.index,}
stat -c %s base64-512MiB.gz.inde

Before: 4839085 B
After : 2463522 B

Note that only half the windows are compressed as of now. To be exact,
the last window in the chunk is not compressed because those are on the
critical path. They should be queued for compression in the future.