[refactor] GzipChunkFetcher: Generalize getter/setters to take generic ChunkData::Configuration defaults
This makes GzipChunkFetcher more agnostic and it is easier to forward
parameters to the decompression tasks. We could now even properly
forward the "statistics enabled" flag!

Window compression was already something that should have been forwarded
like this because the GzipChunkFetcher should not have to know about
any of this.

This also makes ChunkData modifications / subclassing easier to use.
I somehow need to document a use case and example for using the
ChunkData class to do parallelized post-processing with the existing
threading architecture.

In the next commit, we will add Configuration::newlineFormat.