[style] Remove obsolete LUT based precode checks
All of these are strictly worse in every respect. The only thing one
could learn from these is how to handle large sparse LUTs. But, the new
version counting "projected" leaf nodes is simpler and faster and it's
almost embarassing that it took me so long to find that version.